+++ jgclark.Summaries/src/stats.js
@@ -24,7 +24,7 @@ import {
 import { getPeriodStartEndDates, getPeriodStartEndDatesFromPeriodCode } from '@helpers/NPDateTime'
 import { noteOpenInEditor } from '@helpers/NPWindows'
 import { logDebug, logError, logInfo, timer } from '@helpers/dev'
-import { CaseInsensitiveMap, displayTitle, createRunPluginCallbackUrl } from '@helpers/general'
+import { CaseInsensitiveMap, displayTitle, createPrettyRunPluginLink, createRunPluginCallbackUrl } from '@helpers/general'
 import { getOrMakeNote, printNote, replaceSection } from '@helpers/note'
 import { caseInsensitiveCompare } from '@helpers/sorting'
 import { chooseOption, showMessage } from '@helpers/userInput'
@@ -35,20 +35,24 @@ import type { OccurrencesToLookFor } from "./summaryHelpers";
 /**
  * Ask user which period to cover, call main stats function accordingly, and present results
  * @author @jgclark
- * @param {string?} periodCodeArg (optional) lm | mtd | om etc. If not provided user will be asked
- * @param {number?} calNumber (optional)
- * @param {number?} year (optional)
+ * @param {string?} periodCodeArg (optional) lm | mtd | om | YYYY-MM-DD | today etc. If not provided user will be asked.
+ * @param {number?} periodNumberArg (optional)
+ * @param {number?} yearArg (optional)
  */
-export async function statsPeriod(periodCodeArg: string = '', calNumber: number = NaN, year: number = NaN): Promise<void> {
+export async function statsPeriod(periodCodeArg: string = '', periodNumberArg: number = NaN, yearArg: number = NaN): Promise<void> {
   try {
     // Get config from settings
     let config = await getSummariesSettings()
     let fromDate, toDate, periodString, periodShortCode, periodAndPartStr = ''
+    let periodNumber = periodNumberArg
+    let year = yearArg
+    let calendarTimeframe = ''
 
     let isRunningFromXCallback = false
-    if (periodCodeArg && periodCodeArg !== '' && (!isNaN(year) || periodCodeArg === 'today' || new RegExp(`^${RE_DATE}$`).test(periodCodeArg))) {
+    if (periodCodeArg && periodCodeArg !== '' && ((!isNaN(year) && !isNaN(periodNumberArg)) || periodCodeArg === 'today' || new RegExp(`^${RE_DATE}$`).test(periodCodeArg))) {
       isRunningFromXCallback = true
       periodShortCode = periodCodeArg
+      logDebug(pluginJson, `statsPeriod: starting with params ${periodShortCode}, ${periodNumber}, ${year}`)
     }
 
     // Get time period of interest ...
@@ -56,10 +60,19 @@ export async function statsPeriod(periodCodeArg: string = '', calNumber: number 
       // from periodCodeArg
       // TODO: check periodShortCode = week | month | quarter | year | YYYY-MM-DD
       // $FlowIgnore[incompatible-call]
-      [fromDate, toDate, periodShortCode, periodString, periodAndPartStr] = getPeriodStartEndDatesFromPeriodCode(periodShortCode, calNumber, year, config.excludeToday) // note no await
+      [fromDate, toDate, periodShortCode, periodString, periodAndPartStr] = getPeriodStartEndDatesFromPeriodCode(periodShortCode, periodNumber, year, config.excludeToday) // note no await
     } else {
       // or by asking user
-      [fromDate, toDate, periodShortCode, periodString, periodAndPartStr] = await getPeriodStartEndDates('Create stats for which period?', config.excludeToday, '') // note await needed
+      [fromDate, toDate, periodShortCode, periodString, periodAndPartStr, periodNumber] = await getPeriodStartEndDates('Create stats for which period?', config.excludeToday, '') // note await needed
+      year = fromDate.getFullYear()
+
+      calendarTimeframe =
+        (periodShortCode === 'userwtd' || periodShortCode === 'wtd' || periodShortCode === 'lw' || periodShortCode === 'ow') ? 'week'
+          : (periodShortCode === 'mtd' || periodShortCode === 'lm' || periodShortCode === 'om') ? 'month'
+            : (periodShortCode === 'qtd' || periodShortCode === 'lq' || periodShortCode === 'oq') ? 'quarter'
+              : (periodShortCode === 'ytd' || periodShortCode === 'ly' || periodShortCode === 'oy') ? 'year'
+                : 'other'
+      periodShortCode = calendarTimeframe
     }
 
     if (fromDate == null || toDate == null) {
@@ -71,13 +84,7 @@ export async function statsPeriod(periodCodeArg: string = '', calNumber: number 
 
     const fromDateStr = hyphenatedDate(fromDate)
     const toDateStr = hyphenatedDate(toDate)
-    logInfo(pluginJson, `statsPeriod: starting for ${periodString} (${fromDateStr} - ${toDateStr})`)
-    const calendarTimeframe =
-      (periodShortCode === 'userwtd' || periodShortCode === 'wtd' || periodShortCode === 'lw' || periodShortCode === 'ow') ? 'week'
-        : (periodShortCode === 'mtd' || periodShortCode === 'lm' || periodShortCode === 'om') ? 'month'
-          : (periodShortCode === 'qtd' || periodShortCode === 'lq' || periodShortCode === 'oq') ? 'quarter'
-            : (periodShortCode === 'ytd' || periodShortCode === 'ly' || periodShortCode === 'oy') ? 'year'
-              : 'other'
+    logInfo(pluginJson, `statsPeriod: starting with ${periodCodeArg} for ${periodString} (${fromDateStr} - ${toDateStr})`)
 
     let startTime = new Date()
     CommandBar.showLoading(true, `Gathering Data from Calendar notes`)
@@ -101,13 +108,23 @@ export async function statsPeriod(periodCodeArg: string = '', calNumber: number 
       settingsForGO)
     logInfo('statsPeriod', `Gathered all occurrences in ${timer(startTime)}`)
 
-    CommandBar.showLoading(true, `Creating Period Stats`)
-    startTime = new Date()
-    let output = generateProgressUpdate(tmOccurrencesArray, periodString, fromDateStr, toDateStr, 'markdown', config.periodStatsShowSparklines, true).join('\n')
-
-    await CommandBar.onMainThread()
+    // CommandBar.showLoading(true, `Creating Period Stats`)
     CommandBar.showLoading(false)
+    await CommandBar.onMainThread()
+    startTime = new Date()
+    const output = (await generateProgressUpdate(tmOccurrencesArray, periodString, fromDateStr, toDateStr, 'markdown', config.periodStatsShowSparklines, true)).join('\n')
     logInfo('statsPeriod', `Created period stats in ${timer(startTime)}`)
+
+    // Work out a refresh callback string if possible
+    // (namely: if periodNumber known, or a year code, or 'today' or a specific date)
+    let xCallbackMD = ''
+    if (!isNaN(periodNumber) || periodShortCode === 'today' || new RegExp(`^${RE_DATE}$`).test(periodShortCode) || periodShortCode === 'year') {
+      logDebug('periodStats', `Forming refresh callback with params ${periodShortCode}, ${periodNumber}, ${year}`)
+      xCallbackMD =
+        ' ' + createPrettyRunPluginLink('🔄 Refresh', 'jgclark.Summaries', 'periodStats', [periodShortCode, periodNumber, year])
+    } else {
+      logDebug('periodStats', `NOT Forming refresh callback from params ${periodShortCode}, ${periodNumber}, ${year}`)
+    }
 
     // --------------------------------------------------------------------------
     // Ask where to save this summary to, unless isRunningFromXCallback
@@ -139,12 +156,12 @@ export async function statsPeriod(periodCodeArg: string = '', calNumber: number 
         } else {
           // logDebug('statsPeriod', `- about to update section '${config.statsHeading}' in weekly note '${currentNote.filename}' for ${periodAndPartStr}`)
           // add a refresh button if it can work with just a single extra parameter
-          if (periodShortCode.endsWith('td')) {
-            const refreshXCallbackURL = createRunPluginCallbackUrl('jgclark.Summaries', 'periodStats', periodShortCode)
-            output = `[🔄 Refresh](${refreshXCallbackURL})\n${output}`
-          }
+          // if (periodShortCode.endsWith('td')) {
+          // const refreshXCallbackURL = createRunPluginCallbackUrl('jgclark.Summaries', 'periodStats', periodShortCode)
+          // output = `[🔄 Refresh](${refreshXCallbackURL})\n${output}`
+          // }
           // Replace or add output section
-          replaceSection(currentNote, config.statsHeading, `${config.statsHeading} ${periodAndPartStr}`, config.headingLevel, output)
+          replaceSection(currentNote, config.statsHeading, `${config.statsHeading} ${periodAndPartStr}${xCallbackMD}`, config.headingLevel, output)
           logDebug('statsPeriod', `Written results to note '${periodString}'`)
         }
         break
@@ -160,13 +177,13 @@ export async function statsPeriod(periodCodeArg: string = '', calNumber: number 
           // logDebug('statsPeriod', `- about to update section '${config.statsHeading}' in note '${note.filename}' for ${periodAndPartStr}`)
 
           // add a refresh button if it can work with just a single extra parameter
-          if (periodShortCode.endsWith('td')) {
-            const refreshXCallbackURL = createRunPluginCallbackUrl('jgclark.Summaries', 'periodStats', periodShortCode)
-            output = `[🔄 Refresh](${refreshXCallbackURL})\n${output}`
-          }
+          // if (periodShortCode.endsWith('td')) {
+          // const refreshXCallbackURL = createRunPluginCallbackUrl('jgclark.Summaries', 'periodStats', periodShortCode)
+          // output = `[🔄 Refresh](${refreshXCallbackURL})\n${output}`
+          // }
 
           // Replace or add output section
-          replaceSection(note, config.statsHeading, `${config.statsHeading} ${periodAndPartStr}`, config.headingLevel, output)
+          replaceSection(note, config.statsHeading, `${config.statsHeading} ${periodAndPartStr}${xCallbackMD}`, config.headingLevel, output)
           logDebug('statsPeriod', `Written results to note '${periodString}'`)
 
           // Open the results note in a new split window, unless we already have this note open
@@ -204,13 +221,13 @@ export async function statsPeriod(periodCodeArg: string = '', calNumber: number 
           await showMessage(`There was an error getting the Calendar note ${filenameForCalDate} ready to write`)
         } else {
           // add a refresh button if it can work with just a single extra parameter
-          if (periodShortCode.endsWith('td')) {
-            const refreshXCallbackURL = createRunPluginCallbackUrl('jgclark.Summaries', 'periodStats', periodShortCode)
-            output = `[🔄 Refresh](${refreshXCallbackURL})\n${output}`
-          }
+          // if (periodShortCode.endsWith('td')) {
+          //   const refreshXCallbackURL = createRunPluginCallbackUrl('jgclark.Summaries', 'periodStats', periodShortCode)
+          //   output = `[🔄 Refresh](${refreshXCallbackURL})\n${output}`
+          // }
 
           // Replace or add output section
-          replaceSection(note, config.statsHeading, `${config.statsHeading} ${periodAndPartStr}`, config.headingLevel, output)
+          replaceSection(note, config.statsHeading, `${config.statsHeading} ${periodAndPartStr}${xCallbackMD}`, config.headingLevel, output)
           logDebug('statsPeriod', `Written results to note '${periodString}' (filename=${note.filename})`)
           logDebug(pluginJson, `- Editor.note.filename=${note.filename})`)
         }